Create ExportCommentsToExcelTable
This document exports comments from a Word file to an Excel table, with the columns "Heading", "Comment", "Commenter", "Associated Text", and "Page". This macro is intended to help the user consolidate comments in an excel for review.

The "Heading" includes the bullet value.

The "Associated Text" refers to the text highlighted by the Commenter for their comment.

The "Page" is the page number in the heading; it is not the count of pages. So, for example, if the page numbering re-starts for Appendices, then the page number exported to excel will match those re-started values.